Added more ObjectUtils.equals tests and fixed bug with short arrays, #8

+++ core/src/main/java/org/hobsoft/symmetry/util/lang/ObjectUtils.java
@@ -119,7 +119,7 @@ public final class ObjectUtils
 		{
 			equals = Arrays.equals((long[]) a, (long[]) b);
 		}
-		else if (aType == double.class)
+		else if (aType == short.class)
 		{
 			equals = Arrays.equals((short[]) a, (short[]) b);
 		}
